requirements.txt
# Bitcoin Trend Following Strategy Dependencies

# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Data processing and analysis
pandas==2.2.2
numpy==1.26.4

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Data sources
databento==0.32.0
yfinance==0.2.28

# HTTP requests
requests==2.31.0
aiohttp==3.9.1

# Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Logging
loguru==0.7.2

# Background tasks
celery==5.3.4

# Optional: For enhanced features
# redis==5.0.1  # For caching and task queue
# websockets==12.0  # For real-time updates
polygon-api-client==1.14.1
